New filters

See merge request web_project_2018/ShoppingList!126

+++ src/main/java/it/unitn/webprog2018/ueb/shoppinglist/servlets/list/DeleteListServlet.java
@@ -2,7 +2,7 @@
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
-*/
+ */
 package it.unitn.webprog2018.ueb.shoppinglist.servlets.list;
 
 import it.unitn.webprog2018.ueb.shoppinglist.dao.DAOFactory;
@@ -10,6 +10,7 @@ import it.unitn.webprog2018.ueb.shoppinglist.dao.exceptions.DaoException;
 import it.unitn.webprog2018.ueb.shoppinglist.dao.interfaces.ListDAO;
 import it.unitn.webprog2018.ueb.shoppinglist.entities.User;
 import it.unitn.webprog2018.ueb.shoppinglist.entities.List;
+import it.unitn.webprog2018.ueb.shoppinglist.utils.HttpErrorHandler;
 import it.unitn.webprog2018.ueb.shoppinglist.utils.UploadHandler;
 import java.io.IOException;
 import java.util.logging.Level;
@@ -27,17 +28,17 @@ import javax.servlet.http.HttpServletResponse;
  */
 @WebServlet(name = "DeleteListServlet", urlPatterns = {"/restricted/DeleteList/*"})
 public class DeleteListServlet extends HttpServlet {
-	
+
 	private ListDAO listDAO;
-	
+
 	@Inject
 	private UploadHandler uploadHandler;
-	
+
 	@Override
 	public void init() {
 		listDAO = ((DAOFactory) getServletContext().getAttribute("daoFactory")).getListDAO();
 	}
-	
+
 	/**
 	 * Handles the HTTP <code>GET</code> method.
 	 *
@@ -51,21 +52,25 @@ public class DeleteListServlet extends HttpServlet {
 			throws ServletException, IOException {
 		request.setCharacterEncoding("UTF-8");
 		response.setCharacterEncoding("UTF-8");
-		User user =  ((User) request.getSession().getAttribute("user"));
-		try {
-			List list = listDAO.getList((Integer) request.getAttribute("currentListId"));
-			if (listDAO.deleteList(list.getId())) {
-				uploadHandler.deleteFile(list.getImage(), getServletContext());
-				request.getSession().setAttribute("personalLists", listDAO.getPersonalLists(user.getId()));
-				request.getSession().setAttribute("sharedLists", listDAO.getSharedLists(user.getId()));
-				response.sendRedirect(this.getServletContext().getContextPath() + "/restricted/HomePageLogin/" + user.getId());
+		User user = ((User) request.getSession().getAttribute("user"));
+		if ((boolean) request.getAttribute("hasDeletePermission")) {
+			try {
+				List list = (List) request.getAttribute("currentList");
+				if (listDAO.deleteList(list.getId())) {
+					uploadHandler.deleteFile(list.getImage(), getServletContext());
+					request.getSession().setAttribute("personalLists", listDAO.getPersonalLists(user.getId()));
+					request.getSession().setAttribute("sharedLists", listDAO.getSharedLists(user.getId()));
+					response.sendRedirect(this.getServletContext().getContextPath() + "/restricted/HomePageLogin/" + user.getId());
+				}
+			} catch (DaoException ex) {
+				HttpErrorHandler.handleDAOException(ex, response);
+				Logger.getLogger(DeleteListServlet.class.getName()).log(Level.SEVERE, null, ex);
 			}
-		} catch (DaoException ex) {
-			response.sendError(500, "We could not delete your list");
-			Logger.getLogger(DeleteListServlet.class.getName()).log(Level.SEVERE, null, ex);
+		} else {
+			HttpErrorHandler.sendError401(response);
 		}
 	}
-	
+
 	/**
 	 * Handles the HTTP <code>POST</code> method.
 	 *
@@ -78,7 +83,7 @@ public class DeleteListServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 	}
-	
+
 	/**
 	 * Returns a short description of the servlet.
 	 *
@@ -88,5 +93,5 @@ public class DeleteListServlet extends HttpServlet {
 	public String getServletInfo() {
 		return "Deletes a List";
 	}// </editor-fold>
-	
+
 }
